Split hooks into multiple files
